Single subcriber for the transcriber
